requirements.txt
streamlit
requests
google-cloud-storage
pandas
wordcloud
